Allow for local CardType files
Implements #150

+++ modules/RemoteCardType.py
@@ -27,35 +27,42 @@ class RemoteCardType:
         """
         Construct a new RemoteCardType. This downloads the source file at the
         specified location and loads it as a class in the global modules, under
-        the interpreted class name.
+        the interpreted class name. If the given remote specification is a file
+        that exists, that file is loaded.
         
         :param      remote: URL to remote card to inject. Should omit repo base.
                             Should be specified like {username}/{class_name}.
+                            Can also be a local filepath.
         """
 
-        # Make GET request for the contents of the specified value
-        url = f'{self.URL_BASE}/{remote}.py'
-        if (response := get(url)).status_code >= 400:
-            log.error(f'Cannot identify remote Card Type "{remote}"')
-            self.valid = False
-            return None
+        if (file := Path(remote)).exists():
+            # Get class name from file
+            class_name = file.stem
+            file_name = str(file.resolve())
+        else:
+            # Make GET request for the contents of the specified value
+            url = f'{self.URL_BASE}/{remote}.py'
+            if (response := get(url)).status_code >= 400:
+                log.error(f'Cannot identify remote Card Type "{remote}"')
+                self.valid = False
+                return None
 
-        # Succesful request (i.e. file remotely exists)
-        # Get username and class name from the git specification
-        username = remote.split('/')[0]
-        class_name = remote.split('/')[-1]  
+            # Succesful request (i.e. file remotely exists)
+            # Get username and class name from the git specification
+            username = remote.split('/')[0]
+            class_name = remote.split('/')[-1]  
 
-        # Download and write the CardType class into a temporary file
-        file_name = self.TEMP_DIR / f'{username}-{class_name}.py'
-        file_name.parent.mkdir(parents=True, exist_ok=True)
+            # Download and write the CardType class into a temporary file
+            file_name = self.TEMP_DIR / f'{username}-{class_name}.py'
+            file_name.parent.mkdir(parents=True, exist_ok=True)
 
-        # Download file, import as module
-        try:
             # Write remote file contents to temporary class
             with (file_name).open('wb') as fh:
                 fh.write(response.content)
             log.debug(f'Wrote {class_name}.py to {file_name.resolve()}')
 
+        # Import new file as module
+        try:
             # Create module for newly loaded file
             spec = spec_from_file_location(class_name, file_name)
             module = module_from_spec(spec)
@@ -64,11 +71,11 @@ class RemoteCardType:
 
             # Get class from module namespace
             self.card_class = module.__dict__[class_name]
-            log.info(f'Loaded Remote Card Type "{remote}"')
+            log.info(f'Loaded CardType "{remote}"')
             self.valid = True
         except Exception as e:
             # Some error in loading, set object as invalid
-            log.error(f'Cannot load Remote Card Type "{remote}", returned "{e}"')
+            log.error(f'Cannot load CardType "{remote}", returned "{e}"')
             self.card_class = None
             self.valid = False
